Use primary keys in a more sensible way.
Prior to this change, the primary keys were integer values. This was a
little non-sensible as we might as well use the functionality number
as an identifier, rather than having two identifiers (id, functionality
number). This was originally bit messy when it came to handling parent
relationships in the functionality table i.e functionality 1 has 0 as a
parent, 1.1 has 1 as a parent.

This change affects the domain model, controllers and service classes as
well as the insert statements (in import.sql) required to build up the
content of the tables.