Update GUI with basic functionality in place.
This introduces an approach to developing the GUI. We have a
requirements page (kravspec.html) with a non-working progress bar, an
accordian panel menu that loads a description on the right hand side.

This commit also includes an example modal that can be used to show the
various requirements.

While the code builds and retrieves information, there is no useful
functionality available yet