Introduce ProjectRequirement entity
Introducing the ProjectRequirement entity had an effect on other classes
and the domain model. This is reflected in the commit.

For testing purposes requirements2.sql was introduced, but it will be
removed.

The GUI sees continued development and this commit shows for a single
hardcoded project (id=1) some functionality and associated requirements.