Introduce ProjectFunctionality
As per comments in files we want to support that each project has their
own copy of functionality. This is so we save state of progress as the
user progresses through the project.

Although in hindsight, we probably only need state of progress, not a
copy of all the textual values. This will probably be revisited, but
there's no real problem with the duplication. It's just something DB
people hate. Me too ...